load-generator/requirements.txt
requests==2.31.0
oneagent-sdk==1.5.1.20240104.95100
